pyproject.toml
[tool.poetry]
name = "researcher-router"
version = "0.1.0"
description = "A Discord bot that helps researchers find people who discussed specific topics by automatically indexing research papers and discussions using Qdrant vector database."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = ">=1.1.0"
qdrant-client = ">=1.14.2"
discord-py = ">=2.3.2"
sentence-transformers = "^4.1.0"
pydantic = "^2.11.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
